+++ src/main/java/com/github/gtexpert/gtmt/integration/tic/TiCModule.java
@@ -43,4 +43,5 @@ public class TiCModule extends IntegrationSubmodule {
         ElasticMaterialRegistrar.register(event.getRegistry());
         TiCSmeltery.register();
     }
+
 }